requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
twilio==8.9.1
python-multipart==0.0.6
langchain==0.0.329
langchain-openai==0.0.2
chromadb==0.4.15
pydantic==2.4.2
python-dotenv==1.0.0
tiktoken>=0.5.2